+++ src/GUI/Client.java
@@ -11,12 +11,33 @@ package GUI;
 import java.awt.*;
 import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
+import java.io.BufferedReader;
+import java.io.InputStreamReader;
+import java.io.PrintWriter;
+import java.net.Socket;
 
-import javax.swing.ImageIcon;
 import javax.swing.*;
 
 
 public class Client {
+	
+	private String name;
+	private boolean getOut = true;
+	public String serverNum = "localhost";
+	private Label lblServerAddr = new Label ("Server Address:");
+	private Label lblPortNum = new Label ("Port Number:");
+	private Label lblName = new Label ("Name:");
+	private Label lblError = new Label ("Error");
+	private TextField tfName = new TextField();
+	private TextField tfServAddr = new TextField();
+	private TextField tfPortNum = new TextField();
+	private Button bttnConnect = new Button("Connect");
+	private Button bttnCreate = new Button("Create Server");
+	private Button bttnCancel = new Button("Cancel");
+	private int portNum = -1;
+	public BufferedReader readBuff;
+	public PrintWriter writeBuff;
+	
 	private String player1Name = "Player1"; //ConnectUI.name;
 	private String player2Name = "Player2";
 	private String player3Name = "Player3";
@@ -83,7 +104,7 @@ public class Client {
 			}
 		});
 	}
-
+	
 	/**
 	 * Create the application.
 	 */
@@ -91,16 +112,23 @@ public class Client {
 		imagep2 = new ImageIcon(getClass().getResource("/pokeBackp2.jpg"));
 		imagep3 = new ImageIcon(getClass().getResource("/pokeBackp3.jpg"));
 		imagep4 = new ImageIcon(getClass().getResource("/pokeBackp4.jpg"));
-		initialize();
+		start();
 	}
 
 	/**
 	 * Initialize the contents of the frame.
 	 */
-	private void initialize() {
-		//TODO wait until all 4 players are connected
-		
+	
+	private void start() {
 		frame = new JFrame();
+		ConnectUI();
+	}
+	private void initialize() {
+//		//TODO wait until all 4 players are connected
+//		while (numPlayers < 4) {
+//			
+//			
+//		}
 		frame.setResizable(false);
 		frame.getContentPane().setBackground(Color.DARK_GRAY);
 		frame.getContentPane().setForeground(Color.GREEN);
@@ -530,5 +558,131 @@ public class Client {
 			return false;
 		}
 		return false;
+	}
+	
+	public void ConnectUI(){
+		final JFrame ConnectUI = new JFrame();
+		Font f = new Font (lblPortNum.getName(), Font.PLAIN, 15);
+		
+		ConnectUI.setLocation(550, 400);
+		ConnectUI.setTitle("Connect to game");
+		ConnectUI.setLayout(new GridLayout (3, 1));
+		ConnectUI.setSize(500, 130);
+		ConnectUI.setVisible(true);
+		
+		JPanel northPanel = new JPanel (new GridLayout (1, 4));
+		JPanel middlePanel = new JPanel (new GridLayout (1, 2));
+		JPanel southPanel = new JPanel(new GridLayout (1, 4));
+		
+		lblServerAddr.setFont(f);
+		northPanel.add(lblServerAddr);
+		northPanel.add(tfServAddr);
+		
+		lblPortNum.setFont(f);
+		northPanel.add(lblPortNum);
+		northPanel.add(tfPortNum);
+		
+		lblName.setFont(f);
+		middlePanel.add(lblName);
+		
+		lblError.setFont(f);
+		lblError.setForeground(Color.RED);
+		lblError.setVisible(false);
+		southPanel.add(lblError);
+		
+		middlePanel.add(tfName);
+
+		ConnectUI.add(northPanel);
+		ConnectUI.add(middlePanel);
+		
+		southPanel.add(bttnConnect);                    
+		bttnConnect.addActionListener(new ActionListener() {
+            public void actionPerformed(ActionEvent e)
+            {
+            	
+            	
+            	if( !tfPortNum.getText().equals("") && !tfServAddr.getText().equals("") && !tfName.getText().equals("")) {
+            		//serverNum = Integer.parseInt(tfPortNum.getText());
+            		name = tfName.getText();
+            		portNum = Integer.parseInt(tfPortNum.getText());
+            		
+            		//TODO Connect to the server here using portNum and serverNum
+            		try {
+						Socket socket = new Socket(serverNum, portNum);
+						readBuff = new BufferedReader(new InputStreamReader(socket.getInputStream()));
+						writeBuff = new PrintWriter(socket.getOutputStream(), true);
+					} catch (Exception err) {
+						err.printStackTrace();
+	            		lblError.setVisible(true);    
+					}
+            		
+            		
+            		//TODO: Close socket
+            		initialize();
+                	ConnectUI.dispose();
+            		
+            	} else {
+            		lblError.setVisible(true);    
+            	}
+            }
+        });
+		
+		southPanel.add(bttnCreate);                    
+		bttnCreate.addActionListener(new ActionListener() {
+            public void actionPerformed(ActionEvent e)
+            {
+                //TODO create server logic
+            	System.out.println("Need to create server");
+            }
+        });
+		
+		southPanel.add(bttnCancel);                    
+		bttnCancel.addActionListener(new ActionListener() {
+            public void actionPerformed(ActionEvent e)
+            {
+            	ConnectUI.dispose();
+                System.exit(0);
+            }
+        });
+		ConnectUI.add(southPanel);
+		ConnectUI.validate();
+	}
+	
+	public String getPacket(){
+		String myPacket = "";
+		while(true){
+			try{
+				myPacket = readBuff.readLine();
+			}
+			catch (Exception err){
+				System.out.println(err);
+			}
+			if(myPacket.length() > 0)
+				break;
+		}
+
+		return myPacket;
+	}
+	
+	/**
+	 * A function that allows the client to send a packet to the host
+	 * @param packToSend - the packet that needs to be send
+	 */
+	public void sendPacket(String packToSend){
+		while(true){
+			try{
+				// Flush the buffer and send the packet
+				writeBuff.flush();
+				writeBuff.println(packToSend);
+			}
+			catch(Exception err){
+				// catch any errors and print them out
+				System.out.println(err);
+			}
+			// If the writebuffer hasn't encountered any errors, exit the loop
+			//+ otherwise, try to resend the packet
+			if(!writeBuff.checkError())
+				break;
+		}
 	}
 }
